Move FileAccessProxy into its own module
The FileAccessProxy (and its TextFileProxy/BufferedFileProxy/RawFileProxy
subclasses, the io category registrations, and wrap_file_access factory)
were added to collector.py, which was already too large. Extract them into
a dedicated file_access_proxy.py module and import wrap_file_access from
there in the collector.

No behavior change: run_tests.py is still 100% green (339 tests, 1 skip,
0 failures). The regression test's import is updated to point at the new
module.
